chore(kadena-cli) unifying sign
