+++ packages/tools/kadena-cli/src/tx/tests/transaction-live.test.ts
@@ -3,8 +3,8 @@ import { services } from '../../services/index.js';
 import { assertCommandError } from '../../utils/command.util.js';
 import { defaultTemplates } from '../commands/templates/templates.js';
 import { createTransaction } from '../commands/txCreateTransaction.js';
-import { signTransactionFileWithKeyPairAction } from '../commands/txSignWithKeypair.js';
 import { testTransactions } from '../commands/txTestSignedTransaction.js';
+import { signTransactionFileWithKeyPairAction } from '../utils/txSignWithKeypair.js';
 
 const publicKey =
   '2619fafe33b3128f38a4e4aefe6a5559371b18b6c25ac897aff165ce14b241b3';
