+++ packages/tools/kadena-cli/src/tx/txOptions.ts
@@ -167,4 +167,13 @@ export const txOptions = {
       return value;
     },
   }),
+  txSignWith: createOption({
+    key: 'txSignWith',
+    prompt: tx.selectSignMethodPrompt,
+    validation: z.string(),
+    option: new Option(
+      '-s, --tx-sign-with <txSignWith>',
+      'Select a signing method',
+    ),
+  }),
 };
